+++ PLEngine/PLCompositing/src/Shaders/Deferred/SRPDeferredHBAO_Cg.h
@@ -153,7 +153,7 @@ float falloff(float r, float attenuation)\n\
 // Main function\n\
 FS_OUTPUT main(VS_OUTPUT   IN				// Interpolated output from the vertex stage\n\
 	 , uniform int		   NumSteps			// Number of steps (for example 8)\n\
-	 , uniform int		   NumDir			// Number of directions (for example 16)\n\
+	 , uniform float	   NumDir			// Number of directions (for example 16) (should be of integer type, but that's causing troubles on some GPU drivers, see PLCompositing diary entry 02.06.2011 for more details)\n\
 	 , uniform float	   Radius			// Radius of influence (for example 1)\n\
 	 , uniform float	   InvR				// 1/Radius (for example 1)\n\
 	 , uniform float	   SqrR				// Radius*Radius (for example 1)\n\
@@ -218,7 +218,7 @@ FS_OUTPUT main(VS_OUTPUT   IN				// Interpolated output from the vertex stage\n\
 	float alpha = 2 * M_PI / NumDir;\n\
 \n\
 	// Calculate the AO\n\
-	for (int d=0; d<NumDir; d++) {\n\
+	for (int d=0; d<int(NumDir); d++) {\n\
 		float angle = alpha * d;\n\
 		float2 dir = float2(cos(angle), sin(angle));\n\
 		float2 deltaUV = rotate_direction(dir, rand.xy) * step_size.xy;\n\
